requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.9.14
emoji==2.1.0
ephem==4.1.3
PySocks==1.7.1
python-telegram-bot==13.14
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
tornado==6.1
tzdata==2022.2
tzlocal==4.2
